pyproject.toml
[project]
name = "factura-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.3.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "python-dotenv>=1.2.2",
]
